+++ lib/src/element/identification/source.dart
@@ -1,0 +1,14 @@
+import 'package:xml/xml.dart';
+
+import '../../local.dart';
+
+/// https://www.w3.org/2021/06/musicxml40/musicxml-reference/elements/source/
+class Source extends XmlElement {
+  final String value;
+
+  factory Source.parse(XmlElement element) {
+    return Source(element.innerText);
+  }
+
+  Source(this.value) : super.tag(Local.source);
+}